MH-268: (incomplete) Moved component to its own file.
